+++ NEWS.md
@@ -1,5 +1,8 @@
 # ggplot2 2.2.1.9000
 
+* The `show.legend` parameter now accepts a named logical vector to hide/show
+  only some aesthetics in the legend (@tutuchan, #1798)
+
 * Default colour maps for continuous data are controlled by global options
   `ggplot2.continuous.colour` and `ggplot2.continuous.fill`, which can be set to
   either `"gradient"` or `"viridis"` (@karawoo).